+++ frontend/src/app/pages/exercises/exercise/shared/operations-tablet-view/operation-details/operations-vehicles/operations-vehicle-item/operations-vehicle-item.component.ts
@@ -1,7 +1,6 @@
 import { Component, inject, input } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Vehicle } from 'fuesim-digital-shared';
-import { AsyncPipe } from '@angular/common';
 import {
     selectTransferPoints,
     selectAlarmGroups,
@@ -12,7 +11,6 @@ import { AppState } from '../../../../../../../../state/app.state';
     selector: 'app-operations-vehicle-item',
     templateUrl: './operations-vehicle-item.component.html',
     styleUrl: './operations-vehicle-item.component.scss',
-    imports: [AsyncPipe],
 })
 export class OperationsVehicleItemComponent {
     private readonly store = inject(Store<AppState>);
@@ -21,6 +19,7 @@ export class OperationsVehicleItemComponent {
 
     public readonly showTransfer = input<boolean>(true);
 
-    public availableTransferPoints$ = this.store.select(selectTransferPoints);
-    public availableAlarmGroups$ = this.store.select(selectAlarmGroups);
+    public availableTransferPoints =
+        this.store.selectSignal(selectTransferPoints);
+    public availableAlarmGroups = this.store.selectSignal(selectAlarmGroups);
 }