+++ frontend/src/app/pages/exercises/exercise/shared/operations-tablet-view/operational-sections/local-operational-leader/local-operational-leader.component.ts
@@ -1,12 +1,10 @@
 import { Component, inject } from '@angular/core';
-import { createSelector, Store } from '@ngrx/store';
-import { map } from 'rxjs';
-import { AsyncPipe } from '@angular/common';
+import { Store } from '@ngrx/store';
 import { ExerciseService } from '../../../../../../../core/exercise.service';
 import { AppState } from '../../../../../../../state/app.state';
 import {
-    selectVehicles,
-    selectVehiclesInTransfer,
+    selectLocalOperationsCommand,
+    selectVehiclesInTransferFromAlarmgroup,
 } from '../../../../../../../state/application/selectors/exercise.selectors';
 import { SectionLeaderSlotComponent } from '../section-leader-slot/section-leader-slot.component';
 import { OperationsVehicleItemComponent } from '../../operation-details/operations-vehicles/operations-vehicle-item/operations-vehicle-item.component';
@@ -15,23 +13,14 @@ import { OperationsVehicleItemComponent } from '../../operation-details/operatio
     selector: 'app-local-operational-leader',
     templateUrl: './local-operational-leader.component.html',
     styleUrl: './local-operational-leader.component.scss',
-    imports: [
-        SectionLeaderSlotComponent,
-        AsyncPipe,
-        OperationsVehicleItemComponent,
-    ],
+    imports: [SectionLeaderSlotComponent, OperationsVehicleItemComponent],
 })
 export class LocalOperationalLeaderComponent {
     private readonly store = inject(Store<AppState>);
     private readonly exerciseService = inject(ExerciseService);
 
-    public localSectionLeader$ = this.store.select(
-        createSelector(selectVehicles, (vehicles) =>
-            Object.values(vehicles).find(
-                (v) =>
-                    v.operationalAssignment?.type === 'localOperationsCommand'
-            )
-        )
+    public readonly localSectionLeader = this.store.selectSignal(
+        selectLocalOperationsCommand
     );
 
     public onVehicleAssigned(vehicleId: string) {
@@ -44,16 +33,7 @@ export class LocalOperationalLeaderComponent {
         );
     }
 
-    public vehiclesFromAlarmgroups$ = this.store
-        .select(selectVehiclesInTransfer)
-        .pipe(
-            map((vehicles) =>
-                Object.values(vehicles).filter(
-                    (vehicle) =>
-                        vehicle.position.type === 'transfer' &&
-                        vehicle.position.transfer.startPoint.type ===
-                            'alarmGroupStartPoint'
-                )
-            )
-        );
+    public vehiclesFromAlarmgroups = this.store.selectSignal(
+        selectVehiclesInTransferFromAlarmgroup
+    );
 }